Move unit tests back to their place
Some unit tests where moved by mistake along to integration tests inside the starters. Just move them back to their right place.
